Create a method for pagination
It takes the page number, multiplies it by 10 (items per page), which gives the last index necessary for the array slice method to return 10 array items. Then, subtract the last index from the number of items per page to get the start index. For page 1 the start index is always 0.