Fix the page navigation not working with router links
Before, changing pages only updated the browser URL because the code which actually creates the pages was running only on initialization. Now, the "DoCheck" hook compares the page from the URL with the stored page number to rerun page creation if necessary.

Also, now that I'm not using href's anymore, the message logs don't refresh on every page change.

Fixes #1