Prevent the app from throwing users out
Currently users can visit only 2 pages: a users page on the dashboard and the user form. The user form has a cancel button which goes back in the browser history and successfully editing a user's data does the same. Problem is, if the user visits the form from an external link going back in the history will throw them back to where they were before, which should only happen if they use the browser's back button, not an app button.